+++ src/design/plone/contenttypes/testing.py
@@ -8,6 +8,7 @@ from redturtle.volto.testing import RedturtleVoltoRestApiLayer
 from zope.configuration import xmlconfig
 
 import collective.address
+import collective.contentrules.mailfromfield
 import collective.taxonomy
 
 # import collective.folderishtypes
@@ -22,6 +23,7 @@ import plone.app.caching
 import plone.formwidget.geolocation
 import plone.restapi
 import redturtle.bandi
+import redturtle.prenotazioni
 import redturtle.volto
 
 
@@ -74,7 +76,8 @@ class DesignPloneContenttypesRestApiLayer(RedturtleVoltoRestApiLayer):
         self.loadZCML(package=collective.volto.blocksfield)
         self.loadZCML(package=design.plone.contenttypes, context=configurationContext)
         self.loadZCML(package=plone.formwidget.geolocation)
-        self.loadZCML(name="overrides.zcml", package=design.plone.contenttypes)
+        self.loadZCML(package=eea.api.taxonomy)
+        self.loadZCML(package=collective.taxonomy)
         xmlconfig.file(
             "configure.zcml",
             design.plone.contenttypes,
@@ -82,13 +85,19 @@ class DesignPloneContenttypesRestApiLayer(RedturtleVoltoRestApiLayer):
         )
         self.loadZCML(package=redturtle.bandi)
         self.loadZCML(package=kitconcept.seo)
-        self.loadZCML(package=eea.api.taxonomy)
-        self.loadZCML(package=collective.taxonomy)
+
+        # TODO: valutare un layer a parte per i test di redturtle.prenotazioni
+        self.loadZCML(package=redturtle.prenotazioni)
         self.loadZCML(package=collective.z3cform.datagridfield)
+        self.loadZCML(package=collective.contentrules.mailfromfield)
+
+        self.loadZCML(name="overrides.zcml", package=design.plone.contenttypes)
 
     def setUpPloneSite(self, portal):
         super().setUpPloneSite(portal)
         applyProfile(portal, "design.plone.contenttypes:default")
+        # TODO: valutare un layer a parte per i test di redturtle.prenotazioni
+        applyProfile(portal, "redturtle.prenotazioni:default")
 
 
 DESIGN_PLONE_CONTENTTYPES_API_FIXTURE = DesignPloneContenttypesRestApiLayer()